remove whitespaces and commented out code
